requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-core
langchain-groq
sentence-transformers
faiss-cpu
pydantic
streamlit
requests
unstructured
